requirements.txt
streamlit
Pillow
requests
stability-sdk

